+++ Package.swift
@@ -1,5 +1,5 @@
-// swift-tools-version:5.7
-// (Xcode14.0+)
+// swift-tools-version:5.9
+// (Xcode15.0+)
 
 import PackageDescription
 
@@ -9,6 +9,8 @@ let package = Package(
         .iOS(.v13),
         .macOS(.v10_15),
         .macCatalyst(.v14),
+        .visionOS(.v1),
+        .tvOS(.v17),
     ],
     products: [
         .library(
@@ -18,9 +20,9 @@ let package = Package(
     ],
     dependencies: [
         // LK-Prefixed Dynamic WebRTC XCFramework
-        .package(url: "https://github.com/livekit/webrtc-xcframework.git", exact: "125.6422.30"),
-        .package(url: "https://github.com/apple/swift-protobuf.git", from: "1.26.0"),
-        .package(url: "https://github.com/apple/swift-log.git", from: "1.5.4"),
+        .package(url: "https://github.com/livekit/webrtc-xcframework.git", exact: "125.6422.33"),
+        .package(url: "https://github.com/apple/swift-protobuf.git", from: "1.29.0"),
+        .package(url: "https://github.com/apple/swift-log.git", from: "1.6.2"),
         .package(url: "https://github.com/apple/swift-collections.git", from: "1.1.0"),
         // Only used for DocC generation
         .package(url: "https://github.com/apple/swift-docc-plugin.git", from: "1.3.0"),
@@ -47,6 +49,9 @@ let package = Package(
             ],
             resources: [
                 .process("PrivacyInfo.xcprivacy"),
+            ],
+            swiftSettings: [
+                .enableExperimentalFeature("AccessLevelOnImport"),
             ]
         ),
         .testTarget(