+++ Sources/LiveKit/Audio/Manager/AudioManager.swift
@@ -18,17 +18,13 @@ import Accelerate
 import AVFoundation
 import Combine
 
-#if swift(>=5.9)
 internal import LiveKitWebRTC
-#else
-@_implementationOnly import LiveKitWebRTC
-#endif
 
 // Audio Session Configuration related
 public class AudioManager: Loggable {
     // MARK: - Public
 
-    #if compiler(>=6.0)
+    #if swift(>=6.0)
     public nonisolated(unsafe) static let shared = AudioManager()
     #else
     public static let shared = AudioManager()
@@ -58,7 +54,7 @@ public class AudioManager: Loggable {
     ///   - ``isSpeakerOutputPreferred``
     ///
     /// If you want to revert to default behavior, set this to `nil`.
-    @available(*, deprecated, message: "Use `set(engineObservers:)` instead. See `DefaultAudioSessionObserver` for example.")
+    @available(*, deprecated, message: "Use `set(engineObservers:)` instead. See `AudioSessionEngineObserver` for example.")
     public var customConfigureAudioSessionFunc: ConfigureAudioSessionFunc? {
         get { _state.customConfigureFunc }
         set { _state.mutate { $0.customConfigureFunc = newValue } }
@@ -72,8 +68,8 @@ public class AudioManager: Loggable {
     ///
     /// This property is ignored if ``customConfigureAudioSessionFunc`` is set.
     public var isSpeakerOutputPreferred: Bool {
-        get { _state.isSpeakerOutputPreferred }
-        set { _state.mutate { $0.isSpeakerOutputPreferred = newValue } }
+        get { audioSession.isSpeakerOutputPreferred }
+        set { audioSession.isSpeakerOutputPreferred = newValue }
     }
 
     /// Specifies a fixed configuration for the audio session, overriding dynamic adjustments.
@@ -105,16 +101,15 @@ public class AudioManager: Loggable {
         // Keep this var within State so it's protected by UnfairLock
         public var localTracksCount: Int = 0
         public var remoteTracksCount: Int = 0
-        public var isSpeakerOutputPreferred: Bool = true
         public var customConfigureFunc: ConfigureAudioSessionFunc?
         public var sessionConfiguration: AudioSessionConfiguration?
 
         public var trackState: TrackState {
             switch (localTracksCount > 0, remoteTracksCount > 0) {
-            case (true, false): return .localOnly
-            case (false, true): return .remoteOnly
-            case (true, true): return .localAndRemote
-            default: return .none
+            case (true, false): .localOnly
+            case (false, true): .remoteOnly
+            case (true, true): .localAndRemote
+            default: .none
             }
         }
         #endif
@@ -315,7 +310,7 @@ public class AudioManager: Loggable {
     }
 
     /// Set a chain of ``AudioEngineObserver``s.
-    /// Defaults to having a single ``DefaultAudioSessionObserver`` initially.
+    /// Defaults to having a single ``AudioSessionEngineObserver`` initially.
     ///
     /// The first object will be invoked and is responsible for calling the next object.
     /// See ``NextInvokable`` protocol for details.
@@ -324,8 +319,6 @@ public class AudioManager: Loggable {
     public func set(engineObservers: [any AudioEngineObserver]) {
         _state.mutate { $0.engineObservers = engineObservers }
     }
-
-    public let mixer = DefaultMixerAudioObserver()
 
     public var isEngineRunning: Bool {
         RTC.audioDeviceModule.isEngineRunning
@@ -343,6 +336,17 @@ public class AudioManager: Loggable {
         }
     }
 
+    // MARK: - Default AudioEngineObservers
+
+    public let mixer = MixerEngineObserver()
+
+    #if os(iOS) || os(visionOS) || os(tvOS)
+    /// Configures the `AVAudioSession` based on the audio engine's state.
+    /// Set `AudioManager.shared.audioSession.isAutomaticConfigurationEnabled` to `false` to manually configure the `AVAudioSession` instead.
+    /// > Note: It is recommended to set this before connecting to a room.
+    public let audioSession = AudioSessionEngineObserver()
+    #endif
+
     // MARK: - Internal
 
     let _state: StateSync<State>
@@ -351,7 +355,7 @@ public class AudioManager: Loggable {
 
     init() {
         #if os(iOS) || os(visionOS) || os(tvOS)
-        let engineObservers: [any AudioEngineObserver] = [DefaultAudioSessionObserver(), mixer]
+        let engineObservers: [any AudioEngineObserver] = [audioSession, mixer]
         #else
         let engineObservers: [any AudioEngineObserver] = [mixer]
         #endif