+++ Sources/LiveKit/Extensions/RTCMediaConstraints.swift
@@ -16,15 +16,11 @@
 
 import Foundation
 
-#if swift(>=5.9)
 internal import LiveKitWebRTC
-#else
-@_implementationOnly import LiveKitWebRTC
-#endif
 
 extension LKRTCMediaConstraints {
     static let defaultPCConstraints = LKRTCMediaConstraints(
         mandatoryConstraints: nil,
-        optionalConstraints: ["DtlsSrtpKeyAgreement": kRTCMediaConstraintsValueTrue]
+        optionalConstraints: ["DtlsSrtpKeyAgreement": kLKRTCMediaConstraintsValueTrue]
     )
 }