+++ Sources/LiveKit/Track/Remote/RemoteAudioTrack.swift
@@ -17,11 +17,7 @@
 import AVFoundation
 import CoreMedia
 
-#if swift(>=5.9)
 internal import LiveKitWebRTC
-#else
-@_implementationOnly import LiveKitWebRTC
-#endif
 
 @objc
 public class RemoteAudioTrack: Track, RemoteTrack, AudioTrack, @unchecked Sendable {
@@ -49,30 +45,23 @@ public class RemoteAudioTrack: Track, RemoteTrack, AudioTrack, @unchecked Sendab
                    source: source,
                    track: track,
                    reportStatistics: reportStatistics)
-    }
 
-    deinit {
-        // Directly remove the adapter without unnecessary checks
-        guard let audioTrack = mediaTrack as? LKRTCAudioTrack else { return }
-        audioTrack.remove(_adapter)
-    }
-
-    public func add(audioRenderer: AudioRenderer) {
-        let wasEmpty = _adapter.countDelegates == 0
-        _adapter.add(delegate: audioRenderer)
-        // Attach adapter only if it wasn't attached before
-        if wasEmpty {
-            guard let audioTrack = mediaTrack as? LKRTCAudioTrack else { return }
+        if let audioTrack = mediaTrack as? LKRTCAudioTrack {
             audioTrack.add(_adapter)
         }
     }
 
-    public func remove(audioRenderer: AudioRenderer) {
-        _adapter.remove(delegate: audioRenderer)
-        // Remove adapter only if there are no more delegates
-        if _adapter.countDelegates == 0 {
-            guard let audioTrack = mediaTrack as? LKRTCAudioTrack else { return }
+    deinit {
+        if let audioTrack = mediaTrack as? LKRTCAudioTrack {
             audioTrack.remove(_adapter)
         }
+    }
+
+    public func add(audioRenderer: AudioRenderer) {
+        _adapter.add(delegate: audioRenderer)
+    }
+
+    public func remove(audioRenderer: AudioRenderer) {
+        _adapter.remove(delegate: audioRenderer)
     }
 }