+++ Sources/LiveKit/Track/Track.swift
@@ -16,11 +16,7 @@
 
 import Foundation
 
-#if swift(>=5.9)
 internal import LiveKitWebRTC
-#else
-@_implementationOnly import LiveKitWebRTC
-#endif
 
 @objc
 public class Track: NSObject, @unchecked Sendable, Loggable {
@@ -165,10 +161,10 @@ public class Track: NSObject, @unchecked Sendable, Loggable {
             guard let self else { return }
 
             if oldState.dimensions != newState.dimensions {
-                self.log("Track.dimensions \(String(describing: oldState.dimensions)) -> \(String(describing: newState.dimensions))")
+                log("Track.dimensions \(String(describing: oldState.dimensions)) -> \(String(describing: newState.dimensions))")
             }
 
-            self.delegates.notify {
+            delegates.notify {
                 if let delegateInternal = $0 as? TrackDelegateInternal {
                     delegateInternal.track(self, didMutateState: newState, oldState: oldState)
                 }
@@ -177,7 +173,7 @@ public class Track: NSObject, @unchecked Sendable, Loggable {
             if newState.statistics != oldState.statistics || newState.simulcastStatistics != oldState.simulcastStatistics,
                let statistics = newState.statistics
             {
-                self.delegates.notify { $0.track?(self, didUpdateStatistics: statistics, simulcastStatistics: newState.simulcastStatistics) }
+                delegates.notify { $0.track?(self, didUpdateStatistics: statistics, simulcastStatistics: newState.simulcastStatistics) }
             }
         }
     }
@@ -233,13 +229,13 @@ public class Track: NSObject, @unchecked Sendable, Loggable {
     public final func start() async throws {
         try await _startStopSerialRunner.run { [weak self] in
             guard let self else { return }
-            guard self._state.trackState != .started else {
-                self.log("Already started", .warning)
+            guard _state.trackState != .started else {
+                log("Already started", .warning)
                 return
             }
-            try await self.startCapture()
-            if self is RemoteTrack { try await self.enable() }
-            self._state.mutate { $0.trackState = .started }
+            try await startCapture()
+            if self is RemoteTrack { try await enable() }
+            _state.mutate { $0.trackState = .started }
         }
     }
 
@@ -247,13 +243,13 @@ public class Track: NSObject, @unchecked Sendable, Loggable {
     public final func stop() async throws {
         try await _startStopSerialRunner.run { [weak self] in
             guard let self else { return }
-            guard self._state.trackState != .stopped else {
-                self.log("Already stopped", .warning)
+            guard _state.trackState != .stopped else {
+                log("Already stopped", .warning)
                 return
             }
-            try await self.stopCapture()
-            if self is RemoteTrack { try await self.disable() }
-            self._state.mutate { $0.trackState = .stopped }
+            try await stopCapture()
+            if self is RemoteTrack { try await disable() }
+            _state.mutate { $0.trackState = .stopped }
         }
     }