+++ Sources/LiveKit/Types/VideoFrame.swift
@@ -16,11 +16,7 @@
 
 import CoreMedia
 
-#if swift(>=5.9)
 internal import LiveKitWebRTC
-#else
-@_implementationOnly import LiveKitWebRTC
-#endif
 
 public protocol VideoBuffer {}