+++ Sources/LiveKit/Types/VideoRotation.swift
@@ -14,11 +14,7 @@
  * limitations under the License.
  */
 
-#if swift(>=5.9)
 internal import LiveKitWebRTC
-#else
-@_implementationOnly import LiveKitWebRTC
-#endif
 
 public enum VideoRotation: Int, Sendable, Codable {
     case _0 = 0
@@ -27,14 +23,14 @@ public enum VideoRotation: Int, Sendable, Codable {
     case _270 = 270
 }
 
-extension RTCVideoRotation {
+extension LKRTCVideoRotation {
     func toLKType() -> VideoRotation {
         VideoRotation(rawValue: rawValue)!
     }
 }
 
 extension VideoRotation {
-    func toRTCType() -> RTCVideoRotation {
-        RTCVideoRotation(rawValue: rawValue)!
+    func toRTCType() -> LKRTCVideoRotation {
+        LKRTCVideoRotation(rawValue: rawValue)!
     }
 }